Refactor component structure and implement custom hooks for data handling
